requirements.txt
Django==2.1.3
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==3.3
psycopg2==2.7.6.1
PyJWT==1.6.4
python-decouple==3.1
pytz==2018.7
